Upgradle to 4.2 and use GradleTest 1.0-beta12
- Use Torquebox in addition to lasagne,io when building and to work around
  flasky tests.
- Use Grolifant library to deal with Gradle compatibility issues.
- Remove Gradle 5.0 deprecation warning.
- Do not set logging level within tests.
- Ignore IntelliJ "out" directories.
- Update GradleTest fixtures for 1.0-betaX improvements.
- Split out JRubyPluginInstanceSpec to its own file.
- Extract JRubyJarTestKitSpec into its own test file.
- Make test repository preparation a separate task.
- Update TestKit syntax to deal with Gradle 4.2.
- State that Gradle 2.8 is a minimum version.

Also marked custom repository test as `PendingFeature` to be investigated
later as doe a possible rewrite.
