+++ build.gradle
@@ -8,14 +8,15 @@ buildscript {
     dependencies {
         classpath 'com.jfrog.bintray.gradle:gradle-bintray-plugin:1.3.1'
         classpath 'com.netflix.nebula:gradle-extra-configurations-plugin:1.12.+'
-        classpath 'org.ysb33r.gradle:gradletest:0.5.4'
+        classpath 'org.ysb33r.gradle:gradletest:1.0-beta12'
         classpath "org.ajoberstar:gradle-git:1.2.0"
     }
 }
+
 apply plugin: 'idea'
 
 group = 'com.github.jruby-gradle'
-version = '1.5.1'
+version = '1.5.2-SNAPSHOT'
 defaultTasks 'check', 'assemble'
 sourceCompatibility = '1.8'
 targetCompatibility = '1.8'
@@ -25,60 +26,61 @@ if (!releaseBuild) {
 }
 
 subprojects {
-    apply plugin: 'groovy'
-    apply plugin: 'codenarc'
+
     apply plugin: 'idea'
     apply plugin: 'maven'
-    apply plugin: 'com.jfrog.bintray'
-    apply plugin: 'org.ysb33r.gradletest'
-    apply plugin: 'provided-base'
 
     repositories {
         jcenter()
+        maven { url rubygemsProxy }
+        maven { url torqueboxProxy }
     }
 
-    dependencies {
-        compile new GradleDist(project, '2.0').asFileTree
+    if(project.name != 'docs') {
+        apply plugin: 'groovy'
+        apply plugin: 'codenarc'
+        apply plugin: 'com.jfrog.bintray'
+        apply plugin: 'org.ysb33r.gradletest'
+        apply plugin: 'provided-base'
 
-        codenarc(
-            "org.codenarc:CodeNarc:0.24",
-            "org.codehaus.groovy:groovy-all:2.4.0+"
-        )
-    }
+        dependencies {
+//            compile new GradleDist(project, '2.0').asFileTree
 
-    codenarc {
-        sourceSets = [sourceSets.main]
-        configFile = file("${rootProject.projectDir}/gradle/codenarc.rules")
-    }
+            compile localGroovy()
+            compile gradleApi()
+            compile 'org.ysb33r.gradle:grolifant:0.3'
 
-    test {
-        testLogging {
-            showStandardStreams = true
-            exceptionFormat "full"
-        }
-    }
-
-    task sourcesJar(type: Jar, dependsOn: classes) {
-        classifier = 'sources'
-        from sourceSets.main.allSource
-    }
-
-    plugins.withType(JavaPlugin) {
-        sourceCompatibility = 1.8
-        targetCompatibility = 1.8
-
-
-        project.tasks.withType(JavaCompile) { task ->
-            task.sourceCompatibility = project.sourceCompatibility
-            task.targetCompatibility = project.targetCompatibility
+            codenarc(
+                "org.codenarc:CodeNarc:0.24",
+                "org.codehaus.groovy:groovy-all:2.4.0+"
+            )
         }
 
-        project.tasks.withType(GroovyCompile) { task ->
-            task.sourceCompatibility = project.sourceCompatibility
-            task.targetCompatibility = project.targetCompatibility
+        codenarc {
+            sourceSets = [sourceSets.main]
+            configFile = file("${rootProject.projectDir}/gradle/codenarc.rules")
         }
+
+        test {
+            testLogging {
+                showStandardStreams = true
+                exceptionFormat "full"
+            }
+        }
+
+        task sourcesJar(type: Jar, dependsOn: classes) {
+            classifier = 'sources'
+            from sourceSets.main.allSource
+        }
+
+        gradleTest {
+            versions '2.8', '2.14.1', '3.0', '3.5', '4.1'
+            dependsOn jar
+        }
+
+        install.dependsOn check
     }
-    install.dependsOn check
+
 
     idea {
         module {
