+++ examples/run-simple-ruby-script/build.gradle
@@ -3,16 +3,6 @@
  * documentation example. If you change this file, please be sure that it
  * renders appropriately in the generated documentation
  */
-buildscript {
-    repositories { jcenter() }
-
-    dependencies {
-        /* Replace "%%VERSION%%" with the version of JRuby/Gradle you wish to
-         * use if you want to use this build.gradle outside of gradleTest
-         */
-        classpath 'com.github.jruby-gradle:jruby-gradle-plugin:%%VERSION%%'
-    }
-}
 apply plugin: "com.github.jruby-gradle.base"
 
 import com.github.jrubygradle.JRubyExec
