+++ examples/self-executing-jar/build.gradle
@@ -11,12 +11,6 @@ buildscript {
          * GradleTest, this is NOT needed by end-users
          */
         classpath 'com.github.jengelman.gradle.plugins:shadow:[1.2.2,2.0)'
-
-        /* Replace "%%VERSION%%" with the version of JRuby/Gradle you wish to
-         * use if you want to use this build.gradle outside of gradleTest
-         */
-        classpath 'com.github.jruby-gradle:jruby-gradle-plugin:%%VERSION%%'
-        classpath 'com.github.jruby-gradle:jruby-gradle-jar-plugin:%%VERSION%%'
     }
 }
