+++ gradle.properties
@@ -11,3 +11,4 @@ sourceCompatibility=1.7
 jrubyVersion=9.1.12.0
 jettyVersion=9.2.12.v20150709
 rubygemsProxy=http://rubygems.lasagna.io/proxy/maven/releases
+torqueboxProxy=http://rubygems-proxy.torquebox.org/releases
