+++ jruby-gradle-base-plugin/src/gradleTest/jrubyExec/build.gradle
@@ -1,11 +1,5 @@
 import com.github.jrubygradle.JRubyExec
 
-buildscript {
-    dependencies {
-        classpath 'com.github.jruby-gradle:jruby-gradle-plugin:%%VERSION%%'
-    }
-}
-
 apply plugin : 'com.github.jruby-gradle.base'
 
 dependencies {
