+++ jruby-gradle-base-plugin/src/main/groovy/com/github/jrubygradle/internal/JRubyExecTraits.groovy
@@ -9,6 +9,7 @@ import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.Optional
 import org.gradle.internal.FileUtils
 import org.gradle.util.CollectionUtils
+import org.ysb33r.gradle.olifant.StringUtils
 
 /** Provides common traits for JRuby script execution across the {@code JRubyExec}
  * task and {@project.jrubyexec} extension.
@@ -185,7 +186,8 @@ trait JRubyExecTraits {
     Map getPreparedEnvironment(Map env) {
         Map<String, Object> preparedEnv = [:]
 
-        preparedEnv.putAll(env.findAll { String key, Object value ->
+        preparedEnv.putAll(env.findAll { Object rawkey, Object value ->
+            String key = StringUtils.stringize(rawkey)
             inheritRubyEnv || !(key in filterEnvKeys || key.matches(/rvm.*/))
         })
