+++ jruby-gradle-base-plugin/src/test/groovy/com/github/jrubygradle/internal/JRubyExecDelegateSpec.groovy
@@ -28,7 +28,6 @@ class JRubyExecDelegateSpec extends Specification {
     void setup() {
         project = ProjectBuilder.builder().build()
         project.buildDir = TESTROOT
-        project.logging.level = LIFECYCLE
         project.apply plugin: 'com.github.jruby-gradle.base'
     }
