+++ jruby-gradle-jar-plugin/build.gradle
@@ -49,17 +49,14 @@ ext {
     testRepoDir = new File(buildDir,'tmp/test/repo' )
 }
 
+task prepareTestRepo( type : Copy ) {
+    into testRepoDir
+    from project.configurations.testRepo
+    from project.configurations.testRepoOldFiles
+
+}
+
 test {
-    doFirst {
-        copy {
-            from project.configurations.testRepo.files
-            into testRepoDir
-        }
-        copy {
-            from project.configurations.testRepoOldFiles.files
-            into testRepoDir
-        }
-    }
 
     if(gradle.startParameter.isOffline()) {
         systemProperties 'TESTS_ARE_OFFLINE' : '1'
@@ -68,6 +65,8 @@ test {
     systemProperties TESTROOT : new File(buildDir,'tmp/test/unittests').absolutePath
     systemProperties TESTREPO_LOCATION : testRepoDir.absolutePath
     systemProperties 'logback.configurationFile' : new File(projectDir,'src/test/resources/logback-test.xml').absolutePath
+
+    dependsOn prepareTestRepo
 }
 
 
@@ -109,6 +108,17 @@ bintray {
         }
     }
 }
-bintrayUpload.dependsOn assemble
 
+bintrayUpload {
+    dependsOn assemble
+}
+
+gradleTest {
+    dependsOn jar
+    mustRunAfter test, integrationTest
+}
+
+integrationTest {
+    dependsOn prepareTestRepo
+}
 // vim: ft=groovy
