--- jruby-gradle-jar-plugin/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Mon Aug 25 21:52:07 PDT 2014
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-2.0-bin.zip
