+++ jruby-gradle-war-plugin/src/test/groovy/com/github/jrubygradle/war/JRubyWarSpec.groovy
@@ -19,7 +19,6 @@ class JRubyWarSpec extends Specification {
     void setup() {
         project = ProjectBuilder.builder().build()
         project.buildDir = TESTROOT
-        project.logging.level = LIFECYCLE
         project.apply plugin: 'com.github.jruby-gradle.war'
         warTask = project.task(TASK_NAME, type: JRubyWar)
     }
